Add normal notation socket support.
Implement a -s option for all nodejs plugins that takes a socket in the
standard DaZeus notation (starting with unix: or tcp:), and translates it to
the formats used by the -f, -h and -p options. This is particularly useful when
DaZeus auto-starts a plugin, where it inserts unix: or tcp: lines into the
parameters of a plugin.